refactor: directly define agentic AI connectors as beans
