+++ bundle/default-bundle/pom.xml
@@ -131,7 +131,7 @@
     </dependency>
     <dependency>
       <groupId>io.camunda.connector</groupId>
-      <artifactId>connector-agentic-ai-spring-boot-starter</artifactId>
+      <artifactId>connector-agentic-ai</artifactId>
     </dependency>
   </dependencies>
