--- connectors/agentic-ai/agentic-ai-connectors-spring-boot-starter/pom.xml
@@ -1,43 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
-    <parent>
-        <groupId>io.camunda.connector</groupId>
-        <artifactId>connector-agentic-ai-parent</artifactId>
-        <version>8.8.0-SNAPSHOT</version>
-        <relativePath>../pom.xml</relativePath>
-    </parent>
-
-    <name>connector-agentic-ai-spring-boot-starter</name>
-    <description>Agentic AI connectors Spring Boot integration</description>
-    <artifactId>connector-agentic-ai-spring-boot-starter</artifactId>
-    <packaging>jar</packaging>
-
-    <licenses>
-        <license>
-            <name>Camunda Self-Managed Free Edition license</name>
-            <url>
-                https://camunda.com/legal/terms/cloud-terms-and-conditions/camunda-cloud-self-managed-free-edition-terms/
-            </url>
-        </license>
-        <license>
-            <name>Camunda Self-Managed Enterprise Edition license</name>
-        </license>
-    </licenses>
-
-    <dependencies>
-        <dependency>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-autoconfigure</artifactId>
-        </dependency>
-
-        <dependency>
-            <groupId>io.camunda.connector</groupId>
-            <artifactId>connector-agentic-ai</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-    </dependencies>
-</project>
