+++ connectors/agentic-ai/agentic-ai-connectors/src/main/java/io/camunda/connector/agenticai/autoconfigure/AgenticAiConnectorsAutoConfiguration.java
@@ -1,0 +1,67 @@
+/*
+ * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH
+ * under one or more contributor license agreements. Licensed under a proprietary license.
+ * See the License.txt file for more information. You may not use this file
+ * except in compliance with the proprietary license.
+ */
+package io.camunda.connector.agenticai.autoconfigure;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import io.camunda.client.CamundaClient;
+import io.camunda.connector.agenticai.aiagent.AiAgentFunction;
+import io.camunda.connector.agenticai.aiagent.converter.AgentContextMessageSerializer;
+import io.camunda.connector.agenticai.aiagent.provider.ChatModelFactory;
+import io.camunda.connector.agenticai.aiagent.tools.ToolCallingHandler;
+import io.camunda.connector.agenticai.aiagent.tools.ToolSpecificationConverter;
+import io.camunda.connector.agenticai.schema.AdHocToolsSchemaFunction;
+import io.camunda.connector.agenticai.schema.AdHocToolsSchemaResolver;
+import io.camunda.connector.agenticai.schema.CachingAdHocToolsSchemaResolver;
+import io.camunda.connector.agenticai.schema.CamundaClientAdHocToolsSchemaResolver;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@ConditionalOnProperty(
+    value = "camunda.connector.agenticai.enabled",
+    havingValue = "true",
+    matchIfMissing = true)
+@EnableConfigurationProperties(AgenticAiConnectorsConfigurationProperties.class)
+public class AgenticAiConnectorsAutoConfiguration {
+
+  @Bean
+  @ConditionalOnMissingBean
+  public AdHocToolsSchemaResolver adHocToolsSchemaResolver(
+      AgenticAiConnectorsConfigurationProperties configuration, CamundaClient camundaClient) {
+    final var resolver = new CamundaClientAdHocToolsSchemaResolver(camundaClient);
+
+    final var cacheConfiguration = configuration.tools().cache();
+    if (cacheConfiguration.enabled()) {
+      return new CachingAdHocToolsSchemaResolver(
+          resolver,
+          new CachingAdHocToolsSchemaResolver.CacheConfiguration(
+              cacheConfiguration.maxSize(), cacheConfiguration.expireAfterWrite()));
+    }
+
+    return resolver;
+  }
+
+  @Bean
+  @ConditionalOnMissingBean
+  public AiAgentFunction aiAgentFunction(
+      ObjectMapper objectMapper, AdHocToolsSchemaResolver schemaResolver) {
+    return new AiAgentFunction(
+        new ChatModelFactory(),
+        new AgentContextMessageSerializer(objectMapper),
+        new ToolCallingHandler(objectMapper, schemaResolver, new ToolSpecificationConverter()));
+  }
+
+  @Bean
+  @ConditionalOnMissingBean
+  public AdHocToolsSchemaFunction adHocToolsSchemaFunction(
+      AdHocToolsSchemaResolver schemaResolver) {
+    return new AdHocToolsSchemaFunction(schemaResolver);
+  }
+}
