--- connectors/agentic-ai/agentic-ai-connectors/src/main/java/io/camunda/connector/agenticai/core/AgentsApplicationContext.java
@@ -1,16 +1,0 @@
-/*
- * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH
- * under one or more contributor license agreements. Licensed under a proprietary license.
- * See the License.txt file for more information. You may not use this file
- * except in compliance with the proprietary license.
- */
-package io.camunda.connector.agenticai.core;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import io.camunda.connector.agenticai.schema.AdHocToolsSchemaResolver;
-
-public interface AgentsApplicationContext {
-  ObjectMapper objectMapper();
-
-  AdHocToolsSchemaResolver adHocToolSchemaResolver();
-}
