--- connectors/agentic-ai/agentic-ai-connectors/src/main/resources/META-INF/services/io.camunda.connector.api.outbound.OutboundConnectorFunction
@@ -1,2 +1,0 @@
-io.camunda.connector.agenticai.schema.AdHocToolsSchemaFunction
-io.camunda.connector.agenticai.aiagent.AiAgentFunction
