+++ connectors/agentic-ai/agentic-ai-connectors/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,0 +1,1 @@
+io.camunda.connector.agenticai.autoconfigure.AgenticAiConnectorsAutoConfiguration
