+++ connectors/agentic-ai/pom.xml
@@ -18,7 +18,6 @@
 
   <modules>
     <module>agentic-ai-connectors</module>
-    <module>agentic-ai-connectors-spring-boot-starter</module>
   </modules>
 
   <properties>
